pyproject.toml
[project]
name = "botlistbot"
version = "2.0.0"
description = "BotListBot - A Telegram bot list manager"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0,<22.0",
    "pyrogram>=2.0,<3.0",
    "tgcrypto>=1.2",
    "peewee>=3.17",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "logzero>=1.7",
    "inflect>=7.0",
    "python-dateutil>=2.9",
    "requests>=2.31",
    "emoji>=2.0",
    "python-decouple>=3.8",
    "arrow>=1.3",
    "sentry-sdk>=2.0",
    "ordered-set>=4.1",
    "tgintegration>=1.2",
    "minio>=7.2",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
